Cleaning procedures. Package names-> to lower case. Unnecessary comments were removed.
