+++ build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'java'
 apply plugin: 'application'
 apply plugin: 'idea'
 
-mainClassName = "com.company.SpringApplication.Main"
+mainClassName = "com.company.spring_application.Main"
 
 repositories {
     mavenCentral()
