+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'SpringApplication'
-rootProject.name = 'SpringApplication'
+rootProject.name = 'spring_application'
+rootProject.name = 'spring_application'
