+++ src/main/java/com/company/spring_application/publisher/PublisherDashboard.java
@@ -1,4 +1,4 @@
-package com.company.SpringApplication.Publishers;
+package com.company.spring_application.publisher;
 
 /**
  * <p>The PublisherDashboard class is responsible for converting data from <b>JSON<b/> format and
@@ -8,9 +8,7 @@ package com.company.SpringApplication.Publishers;
  * @version 0.1
  */
 public class PublisherDashboard {
-    /**
-     * The default constructor.
-     */
+
     public PublisherDashboard() {
 
     }
