+++ src/main/java/com/company/spring_application/publisher/PublisherDatabase.java
@@ -1,4 +1,4 @@
-package com.company.SpringApplication.Publishers;
+package com.company.spring_application.publisher;
 
 /**
  * <p>The PublisherDatabase class is responsible for converting data from <b>XML<b/> format and
@@ -8,9 +8,7 @@ package com.company.SpringApplication.Publishers;
  * @version 0.1
  */
 public class PublisherDatabase {
-    /**
-     * The default constructor.
-     */
+
     public PublisherDatabase() {
 
     }
