--- src/test/java/MapperXMLTest.java
@@ -1,23 +1,0 @@
-import org.junit.Test;
-
-/**
- * @author AntonKorenev
- * @version 0.1
- */
-public class MapperXMLTest {
-    /**
-     * Method for testing whether incoming format is appropriate or not
-     */
-    @Test
-    public void testInFormat(){
-        System.out.println("Starting 'in' format test");
-    }
-
-    /**
-     * Method for testing whether outgoing format is appropriate or not
-     */
-    @Test
-    public void testOutFormat(){
-        System.out.println("Starting 'out' format test");
-    }
-}
