+++ src/test/java/com/company/spring_application/mapper/MapperXMLTest.java
@@ -1,6 +1,6 @@
-package com.company.SpringApplication.Mappers;
+package com.company.spring_application.mapper;
 
-import com.company.SpringApplication.domain.Order;
+import com.company.spring_application.domain.Order;
 import org.junit.Test;
 import static org.junit.Assert.assertTrue;
 
@@ -9,13 +9,14 @@ public class MapperXMLTest {
     public void conversionOfFileToXMLWasSuccessful() {
         Order testOrder = new Order(-1, "Anton", "Korenev", "Intern", "He wants to test class Order");
         String responseXml = new MapperXML().convert(testOrder);
-        String expectingXml = "<com.company.SpringApplication.domain.Order>\n" +
+        System.out.println(responseXml);
+        String expectingXml = "<com.company.spring__application.domain.Order>\n" +
                 "  <id>-1</id>\n" +
                 "  <firstName>Anton</firstName>\n" +
                 "  <lastName>Korenev</lastName>\n" +
                 "  <position>Intern</position>\n" +
                 "  <taskDescription>He wants to test class Order</taskDescription>\n" +
-                "</com.company.SpringApplication.domain.Order>";
+                "</com.company.spring__application.domain.Order>";
         assertTrue(responseXml.equals(expectingXml));
     }
 }
